# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SubTrackr Subscription Management API
  version: "1.0.0"
  description: |
    The SubTrackr API enables merchants to create and manage subscriptions,
    process payments, configure webhooks, and customise the white-label
    subscription UI for their customers.

    ## Authentication
    All endpoints require a Bearer token in the `Authorization` header.

    ## Versioning
    The current stable version is `v1`. Include `/v1/` in every path.
    Breaking changes will be introduced under `/v2/` with a 6-month deprecation window.

    ## Rate limiting
    Default: **60 requests / minute** per merchant API key.
    Burst allowance: **10 requests / second**.
    Rate-limit headers are returned on every response:
    - `X-RateLimit-Limit`
    - `X-RateLimit-Remaining`
    - `X-RateLimit-Reset` (Unix timestamp)
  contact:
    name: SubTrackr Developer Support
    email: api@subtrackr.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.subtrackr.io/v1
    description: Production
  - url: https://sandbox.subtrackr.io/v1
    description: Sandbox (test environment, no real charges)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: subscription_not_found
        message:
          type: string
          example: No subscription with id sub_abc123 was found.
        details:
          type: object
          additionalProperties: true
...
    BillingCycle:
      type: string
      enum: [monthly, yearly, weekly, custom]
    Plan:
      type: object
      required: [id, name, price, currency, billingCycle]
      properties:
        id:           { type: string, example: plan_monthly_pro }
        name:         { type: string, example: Pro Monthly }
        description:  { type: string, example: Full access with monthly billing }
        price:
          type: number
          format: float
          example: 29.99
        currency:     { type: string, example: USD }
        billingCycle: { $ref: '#/components/schemas/BillingCycle' }
        trialDays:    { type: integer, example: 14, nullable: true }
        features:
          type: array
          items: { type: string }
          example: ["Unlimited projects", "Priority support", "Analytics"]
        metadata:
          type: object
          additionalProperties: true
        createdAt:    { type: string, format: date-time }
        updatedAt:    { type: string, format: date-time }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /plans:
    get:
      tags: [Plans]
      summary: List plans
      parameters:
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Plan' }

    post:
      tags: [Plans]
      summary: Create a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, currency, billingCycle]
              properties:
                name:         { type: string }
                description:  { type: string }
                price:        { type: number }
                currency:     { type: string }
                billingCycle: { $ref: '#/components/schemas/BillingCycle' }
                trialDays:    { type: integer }
                features:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Plan created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
  /plans/{id}:
    get:
      tags: [Plans]
      summary: Retrieve a plan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---- Customers -----------------------------------------------------------
...